feat(report): extract ReportData to model; add ReportDataProvider trait; update querier and generator to use model
