+++ src/report/querier.rs
@@ -5,7 +5,13 @@ use serde::{Deserialize, Serialize};
 use serde_with::{base64::Base64, serde_as};
 use time::OffsetDateTime;
 
-use crate::config::ClickHouseConfig;
+use crate::{
+    config::ClickHouseConfig,
+    report::model::{
+        BigKey, ClusterIssues, DbAggregate, InstanceAggregate, PrefixAggregate, ReportData,
+        TopKeyRecord, TypeAggregate,
+    },
+};
 
 #[serde_as]
 #[derive(Debug, Clone, Serialize)]
@@ -19,52 +25,6 @@ pub struct PrefixRecord {
     pub key_count: u64,
 }
 
-#[serde_as]
-#[derive(Debug, Clone, Serialize)]
-pub struct TopKeyRecord {
-    #[serde_as(as = "Base64")]
-    #[serde(rename = "key_base64")]
-    pub key: Bytes,
-    pub rdb_size: u64,
-    pub member_count: Option<u64>,
-    pub r#type: String,
-    pub instance: String,
-    pub db: u64,
-    pub encoding: String,
-    pub expire_at: Option<String>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct DbAggregate {
-    pub db: u64,
-    pub key_count: u64,
-    pub total_size: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct TypeAggregate {
-    pub data_type: String,
-    pub key_count: u64,
-    pub total_size: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct InstanceAggregate {
-    pub instance: String,
-    pub key_count: u64,
-    pub total_size: u64,
-}
-
-#[serde_as]
-#[derive(Debug, Clone, Serialize)]
-pub struct PrefixAggregate {
-    #[serde_as(as = "Base64")]
-    #[serde(rename = "prefix_base64")]
-    pub prefix: Bytes,
-    pub total_size: u64,
-    pub key_count: u64,
-}
-
 #[derive(Debug)]
 struct PrefixPartition {
     total_size: u64,
@@ -73,36 +33,7 @@ struct PrefixPartition {
     max_key: Bytes,
 }
 
-#[serde_as]
-#[derive(Debug, Clone, Serialize)]
-pub struct BigKey {
-    #[serde_as(as = "Base64")]
-    #[serde(rename = "key_base64")]
-    pub key: Bytes,
-    pub instance: String,
-    pub db: u64,
-    pub r#type: String,
-    pub rdb_size: u64,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ClusterIssues {
-    pub big_keys: Vec<BigKey>,
-    pub codis_slot_skew: bool,
-    pub redis_cluster_slot_skew: bool,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct ReportData {
-    pub cluster: String,
-    pub batch: String,
-    pub db_aggregates: Vec<DbAggregate>,
-    pub type_aggregates: Vec<TypeAggregate>,
-    pub instance_aggregates: Vec<InstanceAggregate>,
-    pub top_keys: Vec<TopKeyRecord>,
-    pub top_prefixes: Vec<PrefixAggregate>,
-    pub cluster_issues: ClusterIssues,
-}
+// Note: `BigKey`, `ClusterIssues`, and `ReportData` types are provided by `crate::report::model`.
 
 pub struct ClickHouseQuerier {
     client: Client,
